move kafka and secrets to their own files
