+++ web3_proxy/src/frontend/authorization.rs
@@ -8,16 +8,17 @@ use crate::caches::RegisteredUserRateLimitKey;
 use crate::errors::{Web3ProxyError, Web3ProxyErrorContext, Web3ProxyResult};
 use crate::globals::{global_db_replica_conn, APP};
 use crate::jsonrpc::{self, JsonRpcId, JsonRpcParams, JsonRpcRequest};
+use crate::kafka::KafkaDebugLogger;
 use crate::response_cache::JsonRpcQueryCacheKey;
 use crate::rpcs::blockchain::Web3ProxyBlock;
 use crate::rpcs::one::Web3Rpc;
+use crate::secrets::RpcSecretKey;
 use crate::stats::{AppStat, BackendRequests};
 use crate::user_token::UserBearerToken;
 use anyhow::Context;
 use axum::headers::authorization::Bearer;
 use axum::headers::{Header, Origin, Referer, UserAgent};
 use chrono::Utc;
-use core::fmt;
 use deferred_rate_limiter::{DeferredRateLimitResult, DeferredRateLimiter};
 use derivative::Derivative;
 use derive_more::From;
@@ -30,12 +31,9 @@ use http::HeaderValue;
 use ipnet::IpNet;
 use migration::sea_orm::prelude::Decimal;
 use migration::sea_orm::{ColumnTrait, EntityTrait, QueryFilter};
-use rdkafka::message::{Header as KafkaHeader, OwnedHeaders as KafkaOwnedHeaders, OwnedMessage};
-use rdkafka::producer::{FutureProducer, FutureRecord};
-use rdkafka::util::Timeout as KafkaTimeout;
 use redis_rate_limiter::redis::AsyncCommands;
 use redis_rate_limiter::{RedisRateLimitResult, RedisRateLimiter};
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::json;
 use serde_json::value::RawValue;
 use std::borrow::Cow;
@@ -44,75 +42,15 @@ use std::fmt::Display;
 use std::hash::{Hash, Hasher};
 use std::mem;
 use std::num::NonZeroU64;
-use std::sync::atomic::{self, AtomicBool, AtomicI64, AtomicU64, AtomicUsize};
+use std::sync::atomic::{self, AtomicBool, AtomicI64, AtomicU64};
 use std::time::Duration;
 use std::{net::IpAddr, str::FromStr, sync::Arc};
 use tokio::sync::RwLock as AsyncRwLock;
 use tokio::sync::{mpsc, OwnedSemaphorePermit, Semaphore};
-use tokio::task::JoinHandle;
 use tokio::time::Instant;
 use tracing::{error, trace, warn};
 use ulid::Ulid;
 use uuid::Uuid;
-
-/// This lets us use UUID and ULID while we transition to only ULIDs
-/// TODO: custom deserialize that can also go from String to Ulid
-#[derive(Copy, Clone, Deserialize)]
-pub enum RpcSecretKey {
-    Ulid(Ulid),
-    Uuid(Uuid),
-}
-
-impl RpcSecretKey {
-    pub fn new() -> Self {
-        Ulid::new().into()
-    }
-
-    fn as_128(&self) -> u128 {
-        match self {
-            Self::Ulid(x) => x.0,
-            Self::Uuid(x) => x.as_u128(),
-        }
-    }
-}
-
-impl PartialEq for RpcSecretKey {
-    fn eq(&self, other: &Self) -> bool {
-        self.as_128() == other.as_128()
-    }
-}
-
-impl Eq for RpcSecretKey {}
-
-impl Debug for RpcSecretKey {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Ulid(x) => Debug::fmt(x, f),
-            Self::Uuid(x) => {
-                let x = Ulid::from(x.as_u128());
-
-                Debug::fmt(&x, f)
-            }
-        }
-    }
-}
-
-/// always serialize as a ULID.
-impl Serialize for RpcSecretKey {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match self {
-            Self::Ulid(x) => x.serialize(serializer),
-            Self::Uuid(x) => {
-                let x: Ulid = x.to_owned().into();
-
-                x.serialize(serializer)
-            }
-        }
-    }
-}
 
 /// TODO: should this have IpAddr and Origin or AuthorizationChecks?
 #[derive(Debug)]
@@ -127,7 +65,7 @@ pub enum RateLimitResult {
     UnknownKey,
 }
 
-#[derive(Clone, Debug)]
+#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
 pub enum AuthorizationType {
     Internal,
     Frontend,
@@ -182,154 +120,12 @@ pub struct Authorization {
     pub authorization_type: AuthorizationType,
 }
 
-pub struct KafkaDebugLogger {
-    topic: String,
-    key: Vec<u8>,
-    headers: KafkaOwnedHeaders,
-    producer: FutureProducer,
-    num_requests: AtomicUsize,
-    num_responses: AtomicUsize,
-}
-
 /// Ulids and Uuids matching the same bits hash the same
 impl Hash for RpcSecretKey {
     fn hash<H: Hasher>(&self, state: &mut H) {
         let x = self.as_128();
 
         x.hash(state);
-    }
-}
-
-impl fmt::Debug for KafkaDebugLogger {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("KafkaDebugLogger")
-            .field("topic", &self.topic)
-            .finish_non_exhaustive()
-    }
-}
-
-type KafkaLogResult = Result<(i32, i64), (rdkafka::error::KafkaError, OwnedMessage)>;
-
-impl KafkaDebugLogger {
-    fn try_new(
-        app: &Web3ProxyApp,
-        authorization: Arc<Authorization>,
-        head_block_num: Option<U64>,
-        kafka_topic: &str,
-        request_ulid: Ulid,
-    ) -> Option<Arc<Self>> {
-        let kafka_producer = app.kafka_producer.clone()?;
-
-        let kafka_topic = kafka_topic.to_string();
-
-        let rpc_secret_key_id = authorization
-            .checks
-            .rpc_secret_key_id
-            .map(|x| x.get())
-            .unwrap_or_default();
-
-        let kafka_key =
-            rmp_serde::to_vec(&rpc_secret_key_id).expect("ids should always serialize with rmp");
-
-        let chain_id = app.config.chain_id;
-
-        let head_block_num = head_block_num.or_else(|| app.balanced_rpcs.head_block_num());
-
-        // TODO: would be nice to have the block hash too
-
-        // another item is added with the response, so initial_capacity is +1 what is needed here
-        let kafka_headers = KafkaOwnedHeaders::new_with_capacity(6)
-            .insert(KafkaHeader {
-                key: "rpc_secret_key_id",
-                value: authorization
-                    .checks
-                    .rpc_secret_key_id
-                    .map(|x| x.to_string())
-                    .as_ref(),
-            })
-            .insert(KafkaHeader {
-                key: "ip",
-                value: Some(&authorization.ip.to_string()),
-            })
-            .insert(KafkaHeader {
-                key: "request_ulid",
-                value: Some(&request_ulid.to_string()),
-            })
-            .insert(KafkaHeader {
-                key: "head_block_num",
-                value: head_block_num.map(|x| x.to_string()).as_ref(),
-            })
-            .insert(KafkaHeader {
-                key: "chain_id",
-                value: Some(&chain_id.to_le_bytes()),
-            });
-
-        // save the key and headers for when we log the response
-        let x = Self {
-            topic: kafka_topic,
-            key: kafka_key,
-            headers: kafka_headers,
-            producer: kafka_producer,
-            num_requests: 0.into(),
-            num_responses: 0.into(),
-        };
-
-        let x = Arc::new(x);
-
-        Some(x)
-    }
-
-    fn background_log(&self, payload: Vec<u8>) -> JoinHandle<KafkaLogResult> {
-        let topic = self.topic.clone();
-        let key = self.key.clone();
-        let producer = self.producer.clone();
-        let headers = self.headers.clone();
-
-        let f = async move {
-            let record = FutureRecord::to(&topic)
-                .key(&key)
-                .payload(&payload)
-                .headers(headers);
-
-            let produce_future =
-                producer.send(record, KafkaTimeout::After(Duration::from_secs(5 * 60)));
-
-            let kafka_response = produce_future.await;
-
-            if let Err((err, msg)) = kafka_response.as_ref() {
-                error!("produce kafka request: {} - {:?}", err, msg);
-                // TODO: re-queue the msg? log somewhere else like a file on disk?
-                // TODO: this is bad and should probably trigger an alarm
-            };
-
-            kafka_response
-        };
-
-        tokio::spawn(f)
-    }
-
-    /// for opt-in debug usage, log the request to kafka
-    /// TODO: generic type for request
-    pub fn log_debug_request(&self, request: &RequestOrMethod) -> JoinHandle<KafkaLogResult> {
-        // TODO: is rust message pack a good choice? try rkyv instead
-        let payload =
-            rmp_serde::to_vec(&request).expect("requests should always serialize with rmp");
-
-        self.num_requests.fetch_add(1, atomic::Ordering::Relaxed);
-
-        self.background_log(payload)
-    }
-
-    pub fn log_debug_response<R>(&self, response: &R) -> JoinHandle<KafkaLogResult>
-    where
-        R: serde::Serialize,
-    {
-        let payload =
-            rmp_serde::to_vec(&response).expect("requests should always serialize with rmp");
-
-        self.num_responses.fetch_add(1, atomic::Ordering::Relaxed);
-
-        self.background_log(payload)
     }
 }
