+++ web3_proxy/src/lib.rs
@@ -16,6 +16,7 @@ pub mod frontend;
 pub mod globals;
 pub mod http_params;
 pub mod jsonrpc;
+pub mod kafka;
 pub mod pagerduty;
 pub mod prelude;
 pub mod premium;
@@ -24,6 +25,7 @@ pub mod referral_code;
 pub mod relational_db;
 pub mod response_cache;
 pub mod rpcs;
+pub mod secrets;
 pub mod stats;
 pub mod test_utils;
 pub mod user_token;
