+++ web3_proxy_cli/src/sub_commands/create_user.rs
@@ -1,5 +1,4 @@
 use tracing::info;
-use web3_proxy::frontend::authorization::RpcSecretKey;
 use web3_proxy::prelude::anyhow::{self, Context};
 use web3_proxy::prelude::argh::{self, FromArgs};
 use web3_proxy::prelude::entities::{rpc_key, user};
@@ -7,6 +6,7 @@ use web3_proxy::prelude::ethers::prelude::Address;
 use web3_proxy::prelude::migration::sea_orm::{self, ActiveModelTrait, TransactionTrait};
 use web3_proxy::prelude::ulid::Ulid;
 use web3_proxy::prelude::uuid::Uuid;
+use web3_proxy::secrets::RpcSecretKey;
 
 #[derive(FromArgs, PartialEq, Debug, Eq)]
 /// Create a new user and api key
