+++ web3_proxy_cli/tests/test_single_proxy.rs
@@ -95,25 +95,21 @@ async fn test_single_proxy_stats_add_up() {
 
     try_join_all(handles).await.unwrap();
 
-    // give stats time to get into the channel
-    // TODO: do this better
-    sleep(Duration::from_secs(5)).await;
+    loop {
+        // give stats time to get into the channel
+        // TODO: do this better
+        sleep(Duration::from_secs(5)).await;
 
-    // Flush all stats here
-    // TODO: the test should maybe pause time so that stats definitely flush from our queries.
-    let flush_0_count_0 = x.flush_stats().await.unwrap();
+        // Flush all stats here
+        // TODO: the test should maybe pause time so that stats definitely flush from our queries.
+        let flush_count = x.flush_stats().await.unwrap();
 
-    warn!("Counts 0 are: {:?}", flush_0_count_0);
-    // we don't actually assert on these because its likely the intervals flushed most of the stats
-    // assert_eq!(flush_0_count_0.relational, 1);
-    // assert_eq!(flush_0_count_0.timeseries, 2);
+        if flush_count.relational + flush_count.timeseries == 0 {
+            break;
+        }
 
-    // Wait a bit. TODO: instead of waiting. make flush stats more robust
-    sleep(Duration::from_secs(5)).await;
-    let flush_0_count_1 = x.flush_stats().await.unwrap();
-    warn!("Counts 0 are: {:?}", flush_0_count_1);
-    assert_eq!(flush_0_count_1.relational, 0);
-    assert_eq!(flush_0_count_1.timeseries, 0);
+        info!(?flush_count);
+    }
 
     // get stats now
     // todo!("Need to validate all the stat accounting now");
